Modify annotations file format
A comment is now on the end of line and separated by ';'. "comment" type is
optional, and an empty (or missing) field also stands for comment.

This provides for fewer typing and also ability to specify a comment on the same
line with data type annotation.

On request by Phire.