Use label-relative if there is a label manually set
If there is an "inner label" (a label inside of instruction, but there is a
manually set label on the instruction itself, prefer that label.

Makes it easier to navigate the code because labels, hidden inside
instructions, are not visible.

This partly reverts commit f35a20cdbc5d00e26b16c5acc309fe0fa9d85372.